lib: move ring_buffer from misc/ to lib/
