+++ lib/ring_buffer/CMakeLists.txt
@@ -1,0 +1,1 @@
+zephyr_sources(ring_buffer.c)
