+++ misc/CMakeLists.txt
@@ -1,3 +1,2 @@
 zephyr_sources_if_kconfig(printk.c)
 zephyr_sources_if_kconfig(reboot.c)
-zephyr_sources_if_kconfig(ring_buffer.c)
