+++ tests/kernel/common/src/ring_buf.c
@@ -7,7 +7,7 @@
  */
 
 #include <ztest.h>
-#include <misc/ring_buffer.h>
+#include <ring_buffer.h>
 #include <logging/sys_log.h>
 
 SYS_RING_BUF_DECLARE_POW2(ring_buf, 8);
